pyproject.toml
[project]
name = "braid-opto-arena"
version = "0.2.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.*"
dependencies = [
    "gitpython>=3.1.44",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "matplotlib>=3.9.3",
    "numba>=0.60.0",
    "numpy>=2.0.2",
    "pandas>=2.2.3",
    "pip>=24.3.1",
    "pre-commit>=4.0.1",
    "pygame>=2.6.1",
    "pyserial>=3.5",
    "pyyaml>=6.0.2",
    "pyzmq>=26.2.0",
    "requests>=2.32.3",
    "ruff>=0.9.1",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.1",
    "tqdm>=4.67.1",
    "ujson>=5.10.0",
]

[dependency-groups]
dev = ["pre-commit>=4.0.1", "ruff>=0.8.2"]
